Move IronMQConstants and change tests to Iron specific package
