+++ src/main/java/org/jenkinsci/plugins/ironmqnotifier/Iron/IronConstants.java
@@ -1,4 +1,4 @@
-package org.jenkinsci.plugins.ironmqnotifier;
+package org.jenkinsci.plugins.ironmqnotifier.Iron;
 
 public final class IronConstants {
