+++ src/main/java/org/jenkinsci/plugins/ironmqnotifier/Iron/IronMQMessage.java
@@ -1,7 +1,6 @@
 package org.jenkinsci.plugins.ironmqnotifier.Iron;
 
 import com.google.gson.Gson;
-import org.jenkinsci.plugins.ironmqnotifier.IronConstants;
 
 import java.util.Date;
