+++ src/main/java/org/jenkinsci/plugins/ironmqnotifier/Iron/IronMessageSettings.java
@@ -1,8 +1,5 @@
 package org.jenkinsci.plugins.ironmqnotifier.Iron;
 
-import org.jenkinsci.plugins.ironmqnotifier.*;
-
-
 
 public class IronMessageSettings {
