+++ src/test/java/org/jenkinsci/plugins/ironmqnotifier/Iron/IronMessageSettingsTest.java
@@ -1,7 +1,6 @@
 package org.jenkinsci.plugins.ironmqnotifier.Iron;
 
 
-import org.jenkinsci.plugins.ironmqnotifier.IronConstants;
 import org.junit.Assert;
 import org.junit.Test;
