+++ src/test/java/org/jenkinsci/plugins/ironmqnotifier/IronConstantsTest.java
@@ -1,5 +1,6 @@
 package org.jenkinsci.plugins.ironmqnotifier;
 
+import org.jenkinsci.plugins.ironmqnotifier.Iron.IronConstants;
 import org.junit.Assert;
 import org.junit.Test;
 import hudson.util.FormValidation;
