+++ src/test/java/org/jenkinsci/plugins/ironmqnotifier/TestSettings.java
@@ -1,6 +1,8 @@
 package org.jenkinsci.plugins.ironmqnotifier;
 
 
+import org.jenkinsci.plugins.ironmqnotifier.Iron.IronConstants;
+
 public class TestSettings {
 
     static String TESTPROJECTID = "141415252";
